move agent service code
